Release 1.1.0:
Add a new option to specify casing convention for routes.
Allow more views/actions to be discovered via [Route] and [HttpGet].
Ignore views that cannot be consumed due to lack of an Action (which includes partials and layouts).
Fix the remaining bugs in link generation.
Add more unit tests (which helped with the bugs!).